Add Categories to ingredients
And show them by categories

+++ resources/views/Ingredients/ingredients.blade.php
@@ -8,29 +8,56 @@
 				<div class="panel-heading">Ingredients <div style="float:right;" id="txt"></div></div>
 
 				<div class="panel-body">
-				<?php
-				if (Session::has('added'))
-				echo '<p style="color:green;">' . Session::get('added') . '</p>';
+					<?php
+					if (Session::has('added'))
+					echo '<p style="color:green;">' . Session::get('added') . '</p>';
 
-				if (Session::has('deleted'))
-				echo '<p style="color:red;">' . Session::get('deleted') . '</p>';
-				?>
-					<h2>Your ingredient !</h2>
-					<?php 
-					foreach ($ingredients as $key => $value) {
-						echo '<p>' . $value->name . ' (' . $value->quantity . ') - <a href="ingredients/delete/' . $value->id . '">Delete</a> - <a href="ingredients/edit/' . $value->id . '">Edit</a></p>';
-					} ?>	
+					if (Session::has('deleted'))
+					echo '<p style="color:red;">' . Session::get('deleted') . '</p>';
 
-					<h2>Add your ingredient !</h2>
-					<form method="post" action="{{ action('IngredientsController@add') }}" accept-charset="UTF-8">
+					if (Session::has('edited'))
+					echo '<p style="color:green;">' . Session::get('edited') . '</p>';
+					?>
+					<button style="float:right;" id="addIngredient">Add your ingredient !</button>
+					<form id="addIngredientForm" method="post" action="{{ action('IngredientsController@add') }}" accept-charset="UTF-8">
 						<input type="text" name="ingredient" placeholder="Name" />
 						<input type="text" name="quantity" placeholder="Quantity" />
+						<select name="ingredientCategory" id="ingredientCategory">
+						<?php 
+							foreach ($ingredientCategories as $key => $value) {
+										echo '<option value="' . $value->id . '">' . $value->name . '</option>';
+									}
+						?>
+						</select>
 						<input type="submit" name="btSubmit" value="Ajouter" />
-					</form>
+					</form>	
+
+					<?php
+					foreach ($ingredientCategories as $value) {
+						$idCat = $value->id;
+						echo '<h3>' . $value->name . '</h3>';
+
+						foreach ($ingredients as $key => $value) {
+							if($value->ingredient_category_id==$idCat)
+							echo '<p>' . $value->name . ' (' . $value->quantity . ') - <a href="ingredients/delete/' . $value->id . '">Delete</a> - <a href="ingredients/edit/' . $value->id . '">Edit</a></p>';
+						}
+						echo '<hr />';
+					}
+
+					//Show ingredients that are not associated with categories
+					echo '<h3>Autres</h3>';
+					foreach ($ingredients as $key => $value) {
+						if($value->ingredient_category_id==0){
+							
+							echo '<p>' . $value->name . ' (' . $value->quantity . ') - <a href="ingredients/delete/' . $value->id . '">Delete</a> - <a href="ingredients/edit/' . $value->id . '">Edit</a></p>';
+						}
+					} ?>
 				</div>
 			</div>
 		</div>
 	</div>
 </div>
+<script src="{{ asset('/js/jquery-1.11.3.js') }}"></script>
 <script src="{{ asset('/js/time.js') }}"></script>
+<script src="{{ asset('/js/main.js') }}"></script>
 @endsection
